+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Unreleased
 
-
-- [added] The `db.reference()` method now optionally takes a `base_url`
+- [added] The `db.reference()` method now optionally takes a `url`
   parameter. This can be used to access multiple Firebase Databases
   in the same project more easily. 
+
+# v2.12.0
+
+- [added] Implemented the ability to create custom tokens without
+  service account credentials.
 - [added] Admin SDK can now read the project ID from both `GCLOUD_PROJECT` and
   `GOOGLE_CLOUD_PROJECT` environment variables.